requirements.txt
# Core
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data preprocessing
imbalanced-learn>=0.11.0
category-encoders>=2.6.0

# Model interpretability
shap>=0.42.0
lime>=0.2.0

# Utilities
jupyter>=1.0.0
notebook>=7.0.0
pandas-profiling>=3.6.0
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0